Log more/rename things
log: Include function name in logs

Improve logging output

+++ src/util.h
@@ -19,7 +19,7 @@ typedef size_t usize;
 /* One second in microseconds */
 #define ANU_TIME_ONE_SEC_IN_US 1000000
 /* Convert microseconds to seconds */
-#define ANU_US_TO_SECONDS(A) ((A) / (ANU_TIME_ONE_SEC_IN_US))
+#define ANU_US_TO_SECONDS(A) ((double)(A) / (ANU_TIME_ONE_SEC_IN_US))
 /* Array size macro */
 #define ANU_ARRAY_SIZE(array) (sizeof(array) / sizeof((array)[0]))
